fix: refactor components + aria label
Fix DataTable's Aria label.

Refactor to a separate component for a cleaner DisplayPointsBalance.tsx.

Move RedeemPointsBalance to a separate child component.

Improve Server's README - add some notes on scaling.
